refactor(home): restructure widgets folder.
