+++ lib/features/home/presentation/views/add_user_view.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:sams_dashboard/core/utils/colors/app_colors.dart';
 import 'package:sams_dashboard/core/utils/styles/app_styles.dart';
-import 'package:sams_dashboard/features/home/presentation/views/widgets/add_user_form_body.dart';
+import 'package:sams_dashboard/features/home/presentation/views/widgets/add_user/add_user_form_body.dart';
 
 class AddUserView extends StatelessWidget {
   const AddUserView({super.key});
