+++ lib/features/home/presentation/views/widgets/users_management_section.dart
@@ -6,9 +6,9 @@ import 'package:sams_dashboard/core/utils/colors/app_colors.dart';
 import 'package:sams_dashboard/core/utils/styles/app_styles.dart';
 import 'package:sams_dashboard/core/widgets/app_animated_loading_indicator.dart';
 import 'package:sams_dashboard/features/home/presentation/view_models/home_cubit/home_cubit.dart';
-import 'package:sams_dashboard/features/home/presentation/views/widgets/table_pagination_bar.dart';
-import 'package:sams_dashboard/features/home/presentation/views/widgets/users_filters_section.dart';
-import 'package:sams_dashboard/features/home/presentation/views/widgets/users_table_section.dart';
+import 'package:sams_dashboard/features/home/presentation/views/widgets/users_table/table_pagination_bar.dart';
+import 'package:sams_dashboard/features/home/presentation/views/widgets/filters/users_filters_section.dart';
+import 'package:sams_dashboard/features/home/presentation/views/widgets/users_table/users_table_section.dart';
 
 //*) Users management main section (Filters + Table + Pagination)
 class UsersManagementSection extends StatelessWidget {
